pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "clamav-sdk"
version = "0.1.0"
description = "Python SDK for the ClamAV API service (REST and gRPC)"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [{ name = "ClamAV SDK Contributors" }]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]
keywords = ["clamav", "antivirus", "malware", "scanning", "grpc", "sdk"]

dependencies = [
    "requests>=2.28.0",
    "grpcio>=1.50.0",
    "protobuf>=4.21.0",
]

[project.optional-dependencies]
async = ["httpx>=0.24.0"]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-asyncio>=0.21.0",
    "responses>=0.23.0",
    "respx>=0.20.0",
    "httpx>=0.24.0",
    "grpcio-tools>=1.50.0",
    "grpcio-testing>=1.50.0",
    "mypy>=1.0",
    "ruff>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/DevHatRo/clamav-api-sdk-python"
Documentation = "https://github.com/DevHatRo/clamav-api-sdk-python#readme"
Repository = "https://github.com/DevHatRo/clamav-api-sdk-python"
Issues = "https://github.com/DevHatRo/clamav-api-sdk-python/issues"

[tool.setuptools.packages.find]
include = ["clamav_sdk*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "integration: marks tests that require a running ClamAV API service (deselect with '-m \"not integration\"')",
]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.coverage.run]
source = ["clamav_sdk"]
omit = ["clamav_sdk/_proto/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "raise$",              # unreachable raise after _handle_rpc_error
    "if TYPE_CHECKING:",
]

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "TCH"]
